requirements.txt
langchain
llama-index
faiss-cpu
chromadb
sentence-transformers
rank-bm25
transformers
torch
pypdf
tqdm
numpy
pandas
scikit-learn
pyyaml
streamlit
ragas
datasets